Persistence v16

+++ Scripts/Persistence/Models/UserGame.cs
@@ -7,8 +7,8 @@ public class UserGame : _Entity{
     public string Code{ get; set; }
 
     public long IdUserGame{ get; set; }
-    public int User{ get; set; }
+    public long User{ get; set; }
     public int Scenario{ get; set; }
-    public int Bank{ get; set; }
+    public long Bank{ get; set; }
     public int PlayTime{ get; set; }
 }
